+++ config.js
@@ -1,55 +1,73 @@
 System.config({
-  "paths": {
-    "*": "*.js",
-    "github:*": "jspm_packages/github/*.js",
-    "aurelia-framework/*": "dist/*.js",
-    "npm:*": "jspm_packages/npm/*.js"
-  }
-});
+  defaultJSExtensions: true,
+  transpiler: "babel",
+  babelOptions: {
+    "optional": [
+      "runtime",
+      "es7.decorators"
+    ]
+  },
+  paths: {
+    "github:*": "jspm_packages/github/*",
+    "aurelia-framework/*": "dist/*",
+    "npm:*": "jspm_packages/npm/*"
+  },
 
-System.config({
-  "map": {
-    "aurelia-binding": "github:aurelia/binding@0.4.0",
-    "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.5.0",
-    "aurelia-loader": "github:aurelia/loader@0.4.0",
-    "aurelia-logging": "github:aurelia/logging@0.2.6",
-    "aurelia-metadata": "github:aurelia/metadata@0.3.4",
-    "aurelia-path": "github:aurelia/path@0.4.6",
-    "aurelia-task-queue": "github:aurelia/task-queue@0.2.5",
-    "aurelia-templating": "github:aurelia/templating@0.9.0",
-    "github:aurelia/binding@0.4.0": {
-      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.5.0",
-      "aurelia-metadata": "github:aurelia/metadata@0.3.4",
-      "aurelia-task-queue": "github:aurelia/task-queue@0.2.5"
+  map: {
+    "aurelia-binding": "github:aurelia/binding@0.9.0",
+    "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.0",
+    "aurelia-loader": "github:aurelia/loader@0.9.0",
+    "aurelia-logging": "github:aurelia/logging@0.7.0",
+    "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+    "aurelia-path": "github:aurelia/path@0.9.0",
+    "aurelia-task-queue": "github:aurelia/task-queue@0.7.0",
+    "aurelia-templating": "github:aurelia/templating@0.15.0",
+    "babel": "npm:babel-core@5.2.2",
+    "babel-runtime": "npm:babel-runtime@5.2.2",
+    "core-js": "npm:core-js@1.1.4",
+    "github:aurelia/binding@0.9.0": {
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.7.0",
+      "core-js": "npm:core-js@0.9.18"
     },
-    "github:aurelia/dependency-injection@0.5.0": {
-      "aurelia-logging": "github:aurelia/logging@0.2.6",
-      "aurelia-metadata": "github:aurelia/metadata@0.3.4",
-      "core-js": "npm:core-js@0.4.10"
+    "github:aurelia/dependency-injection@0.10.0": {
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "core-js": "npm:core-js@0.9.18"
     },
-    "github:aurelia/loader@0.4.0": {
-      "aurelia-html-template-element": "github:aurelia/html-template-element@0.1.3",
-      "aurelia-path": "github:aurelia/path@0.4.6",
-      "core-js": "npm:core-js@0.4.10",
-      "webcomponentsjs": "github:webcomponents/webcomponentsjs@0.5.5"
+    "github:aurelia/loader@0.9.0": {
+      "aurelia-html-template-element": "github:aurelia/html-template-element@0.3.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "core-js": "github:zloirock/core-js@0.8.4"
     },
-    "github:aurelia/templating@0.9.0": {
-      "aurelia-binding": "github:aurelia/binding@0.4.0",
-      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.5.0",
-      "aurelia-html-template-element": "github:aurelia/html-template-element@0.1.3",
-      "aurelia-loader": "github:aurelia/loader@0.4.0",
-      "aurelia-logging": "github:aurelia/logging@0.2.6",
-      "aurelia-metadata": "github:aurelia/metadata@0.3.4",
-      "aurelia-path": "github:aurelia/path@0.4.6",
-      "aurelia-task-queue": "github:aurelia/task-queue@0.2.5",
-      "core-js": "npm:core-js@0.4.10"
+    "github:aurelia/metadata@0.8.0": {
+      "core-js": "npm:core-js@0.9.18"
+    },
+    "github:aurelia/templating@0.15.0": {
+      "aurelia-binding": "github:aurelia/binding@0.9.0",
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.10.0",
+      "aurelia-html-template-element": "github:aurelia/html-template-element@0.3.0",
+      "aurelia-loader": "github:aurelia/loader@0.9.0",
+      "aurelia-logging": "github:aurelia/logging@0.7.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.8.0",
+      "aurelia-path": "github:aurelia/path@0.9.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.7.0",
+      "core-js": "npm:core-js@0.9.18"
     },
     "github:jspm/nodelibs-process@0.1.1": {
       "process": "npm:process@0.10.1"
     },
-    "npm:core-js@0.4.10": {
-      "process": "github:jspm/nodelibs-process@0.1.1"
+    "npm:core-js@0.9.18": {
+      "fs": "github:jspm/nodelibs-fs@0.1.2",
+      "process": "github:jspm/nodelibs-process@0.1.1",
+      "systemjs-json": "github:systemjs/plugin-json@0.1.0"
+    },
+    "npm:core-js@1.1.4": {
+      "fs": "github:jspm/nodelibs-fs@0.1.2",
+      "process": "github:jspm/nodelibs-process@0.1.1",
+      "systemjs-json": "github:systemjs/plugin-json@0.1.0"
     }
   }
 });
-